backend/requirements.txt
# ============================================================
# DualRAG Backend — Production Dependencies (free stack)
# Gemini removed. Claude served via OpenRouter (openai SDK).
# ============================================================

# FastAPI & server
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20

# Pydantic
pydantic==2.11.3
pydantic-settings==2.9.1

# Vector store
qdrant-client==1.14.2

# OpenAI-compatible SDK
# Powers BOTH: nvidia embeddings AND Claude LLM via OpenRouter.
# google-generativeai intentionally removed.
openai==1.82.0

# Document parsing
pdfplumber==0.11.6
python-docx==1.1.2

# HTTP client (NVIDIA reranker)
httpx==0.28.1

# Cloudinary file storage (free tier CDN backup for uploaded docs)
cloudinary==1.41.0

# Environment
python-dotenv==1.1.0
